+++ code/game/objects/items/weapons/weaponry.dm
@@ -21,9 +21,9 @@
 	user.setClickCooldown(DEFAULT_ATTACK_COOLDOWN)
 	user.do_attack_animation(M)
 	//if(user != M)
-	if(user.spell_list.len)
-		user.silence_spells(300) //30 seconds
-		user << "<span class='danger'>You've been silenced!</span>"
+	if(M.spell_list.len)
+		M.silence_spells(300) //30 seconds
+		M << "<span class='danger'>You've been silenced!</span>"
 		return
 
 	if (!(istype(user, /mob/living/carbon/human) || ticker) && ticker.mode.name != "monkey")
@@ -50,6 +50,26 @@
 			M.visible_message("<span class='danger'>\The [user] waves \the [src] over \the [M]'s head.</span>")
 			return
 
+	if(ishuman(M)) //Typecasting, only humans can be vampires
+		var/mob/living/carbon/human/H = M
+
+		if(isvampire(H) && user.mind && (user.mind.assigned_role == "Chaplain")) //Fuck up vampires by smithing the shit out of them. Shock and Awe!
+			if(!(VAMP_SHADOW in H.mind.vampire.powers))
+				H << "\red The [src]'s power violently interferes with your own!"
+				if(H.mind.vampire.nullified < 5) //Don't actually reduce their debuff if it's over 5
+					H.mind.vampire.nullified = max(5, H.mind.vampire.nullified + 2)
+				H.mind.vampire.smitecounter += 30 //Smithe the shit out of him. Four strikes and he's out
+
+/obj/item/weapon/nullrod/pickup(mob/living/user as mob)
+	if(user.mind)
+		if(user.mind.assigned_role == "Chaplain")
+			user << "<span class='notice'>The obsidian rod is teeming with divine power. You feel like you could pulverize a horde of undead with this.</span>"
+		if(ishuman(user)) //Typecasting, only humans can be vampires
+			var/mob/living/carbon/human/H = user
+			if(isvampire(H) && !(VAMP_SHADOW in H.mind.vampire.powers))
+				H.mind.vampire.smitecounter += 60
+				H << "<span class='danger'>You feel an unwanted presence as you pick up the rod. Your body feels like it is burning from the inside!</span>"
+
 /obj/item/weapon/nullrod/afterattack(atom/A, mob/user as mob, proximity)
 	if(!proximity)
 		return
@@ -67,6 +87,7 @@
 	var/net_type = /obj/effect/energy_net
 
 /obj/item/weapon/energy_net/dropped()
+	..()
 	spawn(10)
 		if(src) qdel(src)
 
@@ -82,10 +103,7 @@
 	var/turf/T = get_turf(M)
 	if(T)
 		var/obj/effect/energy_net/net = new net_type(T)
-		net.layer = M.layer+1
-		M.captured = 1
-		net.affecting = M
-		T.visible_message("[M] was caught in an energy net!")
+		net.capture_mob(M)
 		qdel(src)
 
 	// If we miss or hit an obstacle, we still want to delete the net.
@@ -102,75 +120,33 @@
 	opacity = 0
 	mouse_opacity = 1
 	anchored = 1
+	can_buckle = 0 //no manual buckling or unbuckling
 
 	var/health = 25
-	var/mob/living/affecting = null //Who it is currently affecting, if anyone.
-	var/mob/living/master = null    //Who shot web. Will let this person know if the net was successful.
 	var/countdown = -1
 
 /obj/effect/energy_net/teleport
 	countdown = 60
 
-/obj/effect/energy_net/New()
-	..()
-	processing_objects |= src
+/obj/effect/energy_net/proc/capture_mob(mob/living/M)
+	if(M.buckled)
+		M.buckled.unbuckle_mob()
+	buckle_mob(M)
 
-/obj/effect/energy_net/Destroy()
-
-	if(affecting)
-		var/mob/living/carbon/M = affecting
-		M.anchored = initial(affecting.anchored)
-		M.captured = 0
+/obj/effect/energy_net/post_buckle_mob(mob/living/M)
+	if(buckled_mob)
+		layer = M.layer+1
+		visible_message("\The [M] was caught in an energy net!")
+	else
 		M << "You are free of the net!"
-
-	processing_objects -= src
-	..()
+		layer = initial(layer)
+		qdel(src)
 
 /obj/effect/energy_net/proc/healthcheck()
-
 	if(health <=0)
 		density = 0
 		src.visible_message("The energy net is torn apart!")
-		qdel(src)
-	return
-
-/obj/effect/energy_net/process()
-
-	if(isnull(affecting) || affecting.loc != loc)
-		qdel(src)
-		return
-
-	// Countdown begin set to -1 will stop the teleporter from firing.
-	// Clientless mobs can be netted but they will not teleport or decrement the timer.
-	var/mob/living/M = affecting
-	if(countdown == -1 || (istype(M) && !M.client))
-		return
-
-	if(countdown > 0)
-		countdown--
-		return
-
-	// TODO: consider removing or altering this; energy nets are useful on their own
-	// merits and the teleportation was never properly implemented; it's halfassed.
-	density = 0
-	invisibility = 101 //Make the net invisible so all the animations can play out.
-	health = INFINITY  //Make the net invincible so that an explosion/something else won't kill it during anims.
-
-	playsound(affecting.loc, 'sound/effects/sparks4.ogg', 50, 1)
-	anim(affecting.loc,affecting,'icons/mob/mob.dmi',,"phaseout",,affecting.dir)
-
-	affecting.visible_message("[affecting] vanishes in a flare of light!")
-
-	if(holdingfacility.len)
-		affecting.loc = pick(holdingfacility)
-
-	affecting << "You appear in a strange place!"
-
-	playsound(affecting.loc, 'sound/effects/phasein.ogg', 25, 1)
-	playsound(affecting.loc, 'sound/effects/sparks2.ogg', 50, 1)
-	anim(affecting.loc,affecting,'icons/mob/mob.dmi',,"phasein",,affecting.dir)
-
-	qdel(src)
+		unbuckle_mob()
 
 /obj/effect/energy_net/bullet_act(var/obj/item/projectile/Proj)
 	health -= Proj.get_structure_damage()
